+++ R/BIC_AICc_glmnet.R
@@ -10,7 +10,7 @@
 #'
 #' @return A list relevant to model selection.
 #' @author Frederic Bertrand, \email{frederic.bertrand@@math.unistra.fr}
-#' @references \emph{Robust Parametric Classification and Variable Selection by a Minimum Distance Criterion}, Chi and Scott, Journal of Computational and Graphical Statistics, \bold{23}(1), 2014, p111--128, \url{https://doi.org/10.1080/10618600.2012.737296}.
+#' @references \emph{Robust Parametric Classification and Variable Selection by a Minimum Distance Criterion}, Chi and Scott, Journal of Computational and Graphical Statistics, \bold{23}(1), 2014, p111--128, \doi{10.1080/10618600.2012.737296}.
 #' @seealso \code{\link{var_select}}
 #' @examples
 #' set.seed(314)